feat: extract reusable rich text editor
